+++ services/ui-src/src/measures/2022/shared/CommonQuestions/index.tsx
@@ -8,7 +8,7 @@ export * from "./OtherPerformanceMeasure";
 export * from "./Reporting";
 export * from "./StatusOfData";
 export * from "./WhyAreYouNotReporting";
-export * from "shared/commonQuestions/OptionalMeasureStrat";
+export * from "./OptionalMeasureStrat";
 export * from "shared/commonQuestions/DeviationFromMeasureSpecificationCheckboxes";
 export * from "shared/commonQuestions/MeasurementSpecification";
 export * from "./PerformanceMeasure";
