+++ services/ui-src/src/measures/2023/shared/CommonQuestions/OptionalMeasureStrat/context.ts
@@ -5,7 +5,7 @@ import { LabelData } from "utils";
 
 export type ComponentFlagType = "DEFAULT" | "AIF" | "IU" | "PCR";
 
-export interface ContextProps {
+interface ContextProps {
   OPM?: Types.OtherRatesFields[];
   performanceMeasureArray?: Types.RateFields[][];
   IUHHPerformanceMeasureArray?: Types.complexRateFields[][];
