+++ services/ui-src/src/measures/2023/shared/CommonQuestions/OptionalMeasureStrat/omsNodeBuilder.tsx
@@ -5,7 +5,7 @@ import { OmsNode } from "./data";
 
 import { AddAnotherSection } from "./additionalCategory";
 import { SubCatSection } from "./subCatClassification";
-import { NDRSets } from "./NDR/ndrSets";
+import { NDRSets } from "./ndrSets";
 
 interface CheckboxChildrenProps extends OmsNode {
   /** name for react-hook-form registration */
@@ -40,28 +40,20 @@ const renderRadioButtonOptions = ({
   omsNode,
   name,
 }: ChildCheckBoxOptionProps) => {
-  const label = {
-    YesAggregateData: `Yes, we are reporting aggregate data for the ${
-      omsNode?.aggregateTitle || omsNode?.label
-    } categories.`,
-    NoIndependentData: `No, we are reporting disaggregated data for ${
-      omsNode?.aggregateTitle || omsNode?.label
-    } sub-categories`
-  }
-
-  //if year <= 2022
-  const flagSubCat = !!omsNode?.flagSubCat;
-
   return [
     {
-      displayValue: label.YesAggregateData,
+      displayValue: `Yes, we are reporting aggregate data for the ${
+        omsNode?.aggregateTitle || omsNode?.label
+      } categories.`,
       value: "YesAggregateData",
       children: [
-        <NdrNode flagSubCat={flagSubCat} name={name} key={name} />,
+        <NdrNode flagSubCat={!!omsNode?.flagSubCat} name={name} key={name} />,
       ],
     },
     {
-      displayValue: label.NoIndependentData,
+      displayValue: `No, we are reporting disaggregated data for ${
+        omsNode?.aggregateTitle || omsNode?.label
+      } sub-categories`,
       value: "NoIndependentData",
       children: [
         <QMR.Checkbox
